# openapi.yaml
openapi: 3.0.0
info:
    title: Messages API
    version: 1.0.0
servers:
    - url: 'http://localhost:8000'
      description: Local development server
paths:
    /messages:
        get:
            summary: Get Messages
            description: Retrieves a list of messages.
            parameters:
                - name: status
                  in: query
                  description: Filter messages by status
                  required: false
                  schema:
                      type: string
                      enum:
                          - sent
                          - read
                - name: limit
                  in: query
                  description: Number of messages to retrieve
                  required: false
                  schema:
                      type: integer
                      format: int32
                - name: offset
                  in: query
                  description: Offset for pagination
                  required: false
                  schema:
                      type: integer
                      format: int32
                - name: order_by
                  in: query
                  description: Sorting rules for results, e.g., ?order_by[id]=desc
                  required: false
                  style: deepObject
                  explode: true
                  schema:
                      type: object
                      properties:
                          id:
                              type: string
                              enum:
                                  - asc
                                  - desc
                          created_at:
                              type: string
                              enum:
                                  - asc
                                  - desc
            responses:
                '200':
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    messages:
                                        type: array
                                        items:
                                            $ref: '#/components/schemas/Message'
                            example:
                                messages:
                                    - uuid: "123e4567-e89b-12d3-a456-426614174000"
                                      text: "Hello, World!"
                                      status: "read"
                                    - uuid: "123e4567-e89b-12d3-a456-426614174001"
                                      text: "How are you today?"
                                      status: "sent"
                                    - uuid: "123e4567-e89b-12d3-a456-426614174002"
                                      text: "Thanks, I'm fine. How about you?"
                                      status: "sent"
                '400':
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
    /messages/send:
        post:
            summary: Send a Message
            description: The text you want to send
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                text:
                                    type: string
            responses:
                '201':
                    description: Successfully created
                '400':
                    description: Bad Request
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    errors:
                                        type: array
                                        items:
                                            type: string
                '500':
                    description: Internal Server Error
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string

components:
    schemas:
        Message:
            type: object
            properties:
                uuid:
                    type: string
                    format: uuid
                    description: Unique identifier for the message.
                text:
                    type: string
                    description: The content of the message.
                status:
                    type: string
                    enum:
                        - sent
                        - read
                    description: The status of the message.